requirements.txt
asgiref==3.8.1
cmake==3.29.2
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
drf-writable-nested==0.7.0
Pillow==10.4.0
